SPM-830 split Evaluate method to multiple ones
