+++ evaluator/evaluate_advisories.go
@@ -1,0 +1,246 @@
+package evaluator
+
+import (
+	"app/base/database"
+	"app/base/models"
+	"app/base/utils"
+	"github.com/RedHatInsights/patchman-clients/vmaas"
+	"github.com/jinzhu/gorm"
+	"github.com/pkg/errors"
+	"time"
+)
+
+func analyzeAdvisories(tx *gorm.DB, system *models.SystemPlatform, vmaasData *vmaas.UpdatesV2Response) (
+	SystemAdvisoryMap, error) {
+	if !enableAdvisoryAnalysis {
+		utils.Log().Debug("advisory analysis disabled, skipping")
+		return nil, nil
+	}
+
+	patched, unpatched, err := processSystemAdvisories(tx, system, vmaasData, system.InventoryID)
+	if err != nil {
+		evaluationCnt.WithLabelValues("error-process-advisories").Inc()
+		return nil, errors.Wrap(err, "Unable to process system advisories")
+	}
+
+	newSystemAdvisories, err := storeAdvisoryData(tx, system, patched, unpatched)
+	if err != nil {
+		evaluationCnt.WithLabelValues("error-store-advisories").Inc()
+		return nil, errors.Wrap(err, "Unable to store advisory data")
+	}
+	return newSystemAdvisories, nil
+}
+
+// Changes data stored in system_advisories, in order to match newest evaluation
+// Before this methods stores the entries into the system_advisories table, it locks
+// advisory_account_data table, so other evaluations don't interfere with this one
+func processSystemAdvisories(tx *gorm.DB, system *models.SystemPlatform, vmaasData *vmaas.UpdatesV2Response,
+	inventoryID string) (patched []int, unpatched []int, err error) {
+	tStart := time.Now()
+	defer utils.ObserveSecondsSince(tStart, evaluationPartDuration.WithLabelValues("advisories-processing"))
+
+	reported := getReportedAdvisories(vmaasData)
+	oldSystemAdvisories, err := getStoredAdvisoriesMap(tx, system.RhAccountID, system.ID)
+	if err != nil {
+		return nil, nil, errors.Wrap(err, "Unable to get system stored advisories")
+	}
+
+	patched = getPatchedAdvisories(reported, oldSystemAdvisories)
+	updatesCnt.WithLabelValues("patched").Add(float64(len(patched)))
+	utils.Log("inventoryID", inventoryID, "patched", len(patched)).Debug("patched advisories")
+
+	newsAdvisoriesNames, unpatched := getNewAndUnpatchedAdvisories(reported, oldSystemAdvisories)
+	utils.Log("inventoryID", inventoryID, "newAdvisories", len(newsAdvisoriesNames)).Debug("new advisories")
+
+	newIDs, err := getAdvisoriesFromDB(tx, newsAdvisoriesNames)
+	if err != nil {
+		return nil, nil, errors.Wrap(err, "Unable to ensure new system advisories in db")
+	}
+	nUnknown := len(newsAdvisoriesNames) - len(newIDs)
+	if nUnknown > 0 {
+		utils.Log("inventoryID", inventoryID, "unknown", nUnknown).Debug("unknown advisories - ignored")
+		updatesCnt.WithLabelValues("unknown").Add(float64(nUnknown))
+	}
+
+	unpatched = append(unpatched, newIDs...)
+	updatesCnt.WithLabelValues("unpatched").Add(float64(len(unpatched)))
+	utils.Log("inventoryID", inventoryID, "unpatched", len(unpatched)).Debug("patched advisories")
+	return patched, unpatched, nil
+}
+
+func storeAdvisoryData(tx *gorm.DB, system *models.SystemPlatform,
+	patched, unpatched []int) (SystemAdvisoryMap, error) {
+	defer utils.ObserveSecondsSince(time.Now(), evaluationPartDuration.WithLabelValues("advisories-store"))
+	newSystemAdvisories, err := updateSystemAdvisories(tx, system, patched, unpatched)
+	if err != nil {
+		return nil, errors.Wrap(err, "Unable to update system advisories")
+	}
+
+	err = updateAdvisoryAccountDatas(tx, system, patched, unpatched)
+	if err != nil {
+		return nil, errors.Wrap(err, "Unable to update advisory_account_data caches")
+	}
+	return newSystemAdvisories, nil
+}
+
+func getStoredAdvisoriesMap(tx *gorm.DB, accountID, systemID int) (map[string]models.SystemAdvisories, error) {
+	var advisories []models.SystemAdvisories
+	err := database.SystemAdvisoriesBySystemID(tx, accountID, systemID).Preload("Advisory").Find(&advisories).Error
+	if err != nil {
+		return nil, err
+	}
+
+	advisoriesMap := map[string]models.SystemAdvisories{}
+	for _, advisory := range advisories {
+		advisoriesMap[advisory.Advisory.Name] = advisory
+	}
+	return advisoriesMap, nil
+}
+
+func getNewAndUnpatchedAdvisories(reported map[string]bool, stored map[string]models.SystemAdvisories) (
+	[]string, []int) {
+	newAdvisories := []string{}
+	unpatchedAdvisories := []int{}
+	for reportedAdvisory := range reported {
+		if storedAdvisory, found := stored[reportedAdvisory]; found {
+			if storedAdvisory.WhenPatched != nil { // this advisory was already patched and now is un-patched again
+				unpatchedAdvisories = append(unpatchedAdvisories, storedAdvisory.AdvisoryID)
+			}
+		} else {
+			newAdvisories = append(newAdvisories, reportedAdvisory)
+		}
+	}
+	return newAdvisories, unpatchedAdvisories
+}
+
+func getPatchedAdvisories(reported map[string]bool, stored map[string]models.SystemAdvisories) []int {
+	patchedAdvisories := make([]int, 0, len(stored))
+	for storedAdvisory, storedAdvisoryObj := range stored {
+		if _, found := reported[storedAdvisory]; found {
+			continue
+		}
+
+		// advisory contained in reported - it's patched
+		if storedAdvisoryObj.WhenPatched != nil {
+			// it's already marked as patched
+			continue
+		}
+
+		// advisory was patched from last evaluation, let's mark it as patched
+		patchedAdvisories = append(patchedAdvisories, storedAdvisoryObj.AdvisoryID)
+	}
+	return patchedAdvisories
+}
+
+// Return advisory IDs, created advisories count, error
+func getAdvisoriesFromDB(tx *gorm.DB, advisories []string) ([]int, error) {
+	var advisoryIDs []int
+	err := tx.Model(&models.AdvisoryMetadata{}).Where("name IN (?)", advisories).
+		Pluck("id", &advisoryIDs).Error
+	if err != nil {
+		return nil, err
+	}
+	return advisoryIDs, nil
+}
+
+func ensureSystemAdvisories(tx *gorm.DB, rhAccountID int, systemID int, advisoryIDs []int) error {
+	advisoriesObjs := models.SystemAdvisoriesSlice{}
+	for _, advisoryID := range advisoryIDs {
+		advisoriesObjs = append(advisoriesObjs,
+			models.SystemAdvisories{RhAccountID: rhAccountID, SystemID: systemID, AdvisoryID: advisoryID})
+	}
+
+	txOnConflict := tx.Set("gorm:insert_option", "ON CONFLICT DO NOTHING")
+	err := database.BulkInsert(txOnConflict, advisoriesObjs)
+	return err
+}
+
+func lockAdvisoryAccountData(tx *gorm.DB, system *models.SystemPlatform, patched, unpatched []int) error {
+	// Lock advisory-account data, so it's not changed by other concurrent queries
+	var aads []models.AdvisoryAccountData
+	err := tx.Set("gorm:query_option", "FOR UPDATE OF advisory_account_data").
+		Order("advisory_id").
+		Find(&aads, "rh_account_id = ? AND (advisory_id in (?) OR advisory_id in (?))",
+			system.RhAccountID, patched, unpatched).Error
+
+	return err
+}
+
+func calcAdvisoryChanges(system *models.SystemPlatform, patched, unpatched []int) []models.AdvisoryAccountData {
+	aadMap := map[int]models.AdvisoryAccountData{}
+	// If system is stale, we won't change any rows  in advisory_account_data
+	if system.Stale {
+		return []models.AdvisoryAccountData{}
+	}
+
+	for _, id := range unpatched {
+		aadMap[id] = models.AdvisoryAccountData{
+			AdvisoryID:      id,
+			RhAccountID:     system.RhAccountID,
+			SystemsAffected: 1,
+		}
+	}
+
+	for _, id := range patched {
+		aadMap[id] = models.AdvisoryAccountData{
+			AdvisoryID:      id,
+			RhAccountID:     system.RhAccountID,
+			SystemsAffected: -1,
+		}
+	}
+
+	deltas := make([]models.AdvisoryAccountData, 0, len(patched)+len(unpatched))
+	for _, aad := range aadMap {
+		deltas = append(deltas, aad)
+	}
+	return deltas
+}
+
+func deleteOldSystemAdvisories(tx *gorm.DB, accountID, systemID int, patched []int) error {
+	err := tx.Where("rh_account_id = ? ", accountID).
+		Where("system_id = ?", systemID).
+		Where("when_patched IS NOT NULL or advisory_id in (?)", patched).
+		Delete(&models.SystemAdvisories{}).Error
+	return err
+}
+
+func updateSystemAdvisories(tx *gorm.DB, system *models.SystemPlatform,
+	patched, unpatched []int) (SystemAdvisoryMap, error) {
+	// this will remove many many old items from our "system_advisories" table
+	err := deleteOldSystemAdvisories(tx, system.RhAccountID, system.ID, patched)
+	if err != nil {
+		return nil, err
+	}
+
+	err = ensureSystemAdvisories(tx, system.RhAccountID, system.ID, unpatched)
+	if err != nil {
+		return nil, err
+	}
+
+	newSystemAdvisories := SystemAdvisoryMap{}
+	var data []models.SystemAdvisories
+	err = tx.Preload("Advisory").
+		Find(&data, "system_id = ? AND rh_account_id = ?", system.ID, system.RhAccountID).Error
+
+	if err != nil {
+		return nil, err
+	}
+
+	for _, sa := range data {
+		newSystemAdvisories[sa.Advisory.Name] = sa
+	}
+	return newSystemAdvisories, nil
+}
+
+func updateAdvisoryAccountDatas(tx *gorm.DB, system *models.SystemPlatform, patched, unpatched []int) error {
+	err := lockAdvisoryAccountData(tx, system, patched, unpatched)
+	if err != nil {
+		return err
+	}
+
+	changes := calcAdvisoryChanges(system, patched, unpatched)
+	txOnConflict := database.OnConflictDoUpdateExpr(tx, []string{"rh_account_id", "advisory_id"},
+		database.UpExpr{Name: "systems_affected", Expr: "advisory_account_data.systems_affected + excluded.systems_affected"})
+
+	return database.BulkInsert(txOnConflict, changes)
+}
