+++ evaluator/remediations_test.go
@@ -31,7 +31,7 @@ var testVmaasResponse = vmaas.UpdatesV2Response{
 
 func TestCreateRemediationsState(t *testing.T) {
 	id := "00000000-0000-0000-0000-000000000012"
-	state := createRemediationsStateMsg(id, testVmaasResponse)
+	state := createRemediationsStateMsg(id, &testVmaasResponse)
 	assert.NotNil(t, state)
 	assert.Equal(t, state.HostID, id)
 	assert.Equal(t, state.Issues, []string{"patch:RH-1", "patch:RH-100", "patch:RH-2",
